Removed javascript file since dart is using a new naming convention for generated javascript. 	deleted:    examples/canvas_geometry_cube/Canvas_Geometry_Cube.dart.app.js
New javascript file
	new file:   examples/canvas_geometry_cube/Canvas_Geometry_Cube.dart.js

Updated the example to point to the correct file
	modified:   examples/canvas_geometry_cube/Canvas_Geometry_Cube.html

Adding script to build with minfrog.
	new file:   examples/canvas_geometry_cube/buildit.sh

minfrog did not want to compile with the duplicated #source file.
Consider moving each src folder into a lib.
	modified:   src/ThreeD.dart

Using default values for the optional parameters in the constructor. Removed
the call to SetValue
	modified:   src/core/Quaternion.dart

Using default values for the optional parameters in the constructor
	modified:   src/core/Vector3.dart

Removed the private memeber because of compile time issues. CubeGeometry
extends Geometry which is conflicting having two private members with
the same name. I think this might be a bug, would have to check the
spec. http://try.dartlang.org/s/H4Ur seems to be fine with that..
	modified:   src/extras/geometries/CubeGeometry.dart

Same issue as CubeGeometry
	modified:   src/scenes/Scene.dart